requirements.txt
requests
dotenv
flask
faiss-cpu==1.7.4
numpy